requirements.txt
ansible-compat==2.2.4
ansible-core==2.13.5
ansible-lint==6.8.6
arrow==1.2.3
attrs==22.1.0
binaryornot==0.4.4
black==22.10.0
bracex==2.3.post1
certifi==2022.9.24
cffi==1.15.1
cfgv==3.3.1
chardet==5.0.0
charset-normalizer==2.1.1
check-jsonschema==0.18.4
click==8.1.3
click-help-colors==0.9.1
commonmark==0.9.1
cookiecutter==2.1.1
cryptography==38.0.3
distlib==0.3.6
distro==1.8.0
docker==6.0.1
enrich==1.2.7
exceptiongroup==1.0.0
filelock==3.8.0
flake8==5.0.4
identify==2.5.8
idna==3.4
iniconfig==1.1.1
Jinja2==3.1.2
jinja2-time==0.2.0
jsonschema==4.17.0
MarkupSafe==2.1.1
mccabe==0.7.0
molecule==4.0.3
molecule-docker==2.1.0
msgpack==1.0.4
mypy-extensions==0.4.3
nodeenv==1.7.0
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==7.2.0
pytest-mock==3.10.0
python-dateutil==2.8.2
python-slugify==6.1.2
PyYAML==6.0
requests==2.28.1
resolvelib==0.8.1
rich==12.6.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
selinux==0.2.1
six==1.16.0
subprocess-tee==0.3.5
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
urllib3==1.26.12
virtualenv==20.16.6
wcmatch==8.4.1
websocket-client==1.4.1
yamllint==1.28.0
